backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
requests